+++ database/migrations/2024_10_31_190227_create_kondisi_tanah_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('id_kriteria')->constrained('data_kriteria');
             $table->foreignId('id_tanah')->constrained('data_tanah');
-            $table->integer('nilai');
+            $table->double('nilai');
             $table->string('bulan');
             $table->timestamps();
         });
